requirements.txt
python-dotenv
openai
flask
flask-cors
PyPDF2
python-docx
asyncio
aiofiles
concurrent.futures 